refactor: move assignment of global device potential variables to the same file as they are defined
we move the data copying (via `hipMemcpyToSymbol`) of global device variables for potential metadata
and potential splines to the same fale as they are defined.

This change is a workaround to fix a bug of the linkage of shadow variable in clang.
(see also https://github.com/ROCm-Developer-Tools/HIP/issues/2249#issuecomment-810322623)